+++ packages/cli-forger/src/index.ts
@@ -1,3 +1,4 @@
-export * from './main'
-export * from './command'
-export * from './option'
+export * from "./main";
+export * from "./command";
+export * from "./option";
+export { type Workers, type WorkersExtensions } from "./workers";