+++ packages/cli-forger/src/option.ts
@@ -1,97 +1,99 @@
-import { ValueObject } from './base/valueObject'
-import { Optional } from './types/optional'
+import { ValueObject } from "./base/valueObject";
+import { Optional } from "./types/optional";
 
 export enum OptionType {
-  STRING = 'string',
-  LIST = 'list',
-  BOOLEAN = 'boolean',
+  STRING = "string",
+  LIST = "list",
+  BOOLEAN = "boolean",
 }
 
-type OptionTypeRaw = 'list' | 'string' | 'boolean'
+type OptionTypeRaw = "list" | "string" | "boolean";
 
 interface OptionProps {
-  name: string
-  description: string
-  long: string
-  type: OptionType | OptionTypeRaw
-  defaultValue: string | boolean | null | undefined
-  short: string | null
-  required: boolean
-  handler: ((...args: unknown[]) => void) | null
+  name: string;
+  description: string;
+  long: string;
+  type: OptionType | OptionTypeRaw;
+  defaultValue: string | boolean | null | undefined;
+  short: string | null;
+  required: boolean;
+  handler: ((...args: unknown[]) => void) | null;
 }
 
 export class Option extends ValueObject<OptionProps> {
   static create(
     props: Optional<
       OptionProps,
-      'short' | 'required' | 'long' | 'defaultValue' | 'type' | 'handler'
+      "short" | "required" | "long" | "defaultValue" | "type" | "handler"
     >,
   ) {
     const optionProps: OptionProps = {
       short: props.short ?? null,
       required: props.required ?? false,
-      long: props.long?.toLowerCase() ?? props.name.toLowerCase(),
+      long:
+        props.long?.toLowerCase() ??
+        props.name.toLowerCase().replace(/\s+/g, "-"),
       defaultValue: props.defaultValue ?? undefined,
-      type: props.type ?? 'string',
+      type: props.type ?? "string",
       handler: props.handler ?? null,
       ...props,
-    }
+    };
 
-    const options = new Option(optionProps)
-    return options
+    const options = new Option(optionProps);
+    return options;
   }
 
   getFlags() {
-    let flags = ''
+    let flags = "";
 
     if (this.short) {
-      flags = `-${this.short}, `
+      flags = `-${this.short}, `;
     }
 
-    flags += `--${this.long} `
+    flags += `--${this.long} `;
 
     if (this.type !== OptionType.BOOLEAN) {
       if (this.required) {
-        flags += `<${this.long.toUpperCase()}${this.type === 'list' ? '...' : ''}>`
+        flags += `<${this.long.toUpperCase()}${this.type === "list" ? "..." : ""}>`;
       }
 
       if (!this.required) {
-        flags += `[${this.long.toUpperCase()}${this.type === 'list' ? '...' : ''}]`
+        flags += `[${this.long.toUpperCase()}${this.type === "list" ? "..." : ""}]`;
       }
     }
 
-    return flags
+    return flags;
   }
 
   get name() {
-    return this.props.name
+    return this.props.name;
   }
 
   get description() {
-    return this.props.description
+    return this.props.description;
   }
 
   get long() {
-    return this.props.long
+    return this.props.long;
   }
 
   get short() {
-    return this.props.short
+    return this.props.short;
   }
 
   get required() {
-    return this.props.required
+    return this.props.required;
   }
 
   get defaultValue() {
-    return this.props.defaultValue
+    return this.props.defaultValue;
   }
 
   get type() {
-    return this.props.type
+    return this.props.type;
   }
 
   get handler() {
-    return this.props.handler
+    return this.props.handler;
   }
 }