+++ packages/cli-forger/src/workers/files/create.ts
@@ -3,18 +3,50 @@ import { workers } from "../index";
 import { File } from "./File";
 import { isString } from "lodash";
 
+interface CreateOptions {
+  exitOnExists?: boolean;
+  updateOnExists?: boolean;
+  messageWhenExists?: string;
+  showInfosLog?: boolean;
+}
+
 export function create(
   rawPaths: string[] | string,
   content: { [x: number | string]: any } | string,
+  options: CreateOptions = {},
 ) {
-  const path = workers.path.getPath(rawPaths);
+  const {
+    showInfosLog = false,
+    exitOnExists = true,
+    updateOnExists = false,
+    messageWhenExists = " File already exists!",
+  } = options;
 
-  if (workers.files.exists(path)) {
+  const path = workers.path.getPath(rawPaths);
+  const fileExists = workers.files.exists(path);
+
+  if (fileExists && exitOnExists) {
     workers.logger.exit.error(
-      "File already exists!",
+      messageWhenExists,
       `files.create ${workers.figures.pointer} ${cwd()}`,
       ` create.path ${workers.figures.pointer} ${path}`,
     );
+  }
+
+  if (fileExists) {
+    const file = workers.files.read(path);
+
+    if (updateOnExists) {
+      if (showInfosLog) {
+        workers.logger.info(` File ${path} already exists`);
+        workers.logger.info(` Updatind`);
+      }
+
+      file.set(content);
+      file.save();
+    }
+
+    return file;
   }
 
   const file = File.create({
@@ -22,7 +54,10 @@ export function create(
     rawContent: isString(content) ? content : JSON.stringify(content, null, 2),
   });
 
-  workers.logger.info(` Creating file ${workers.figures.pointer} ${path}`);
+  if (showInfosLog) {
+    workers.logger.info(` Creating file ${workers.figures.pointer} ${path}`);
+  }
+
   file.save();
 
   return file;