+++ packages/cli-forger/src/workers/folders/index.ts
@@ -1,2 +1,4 @@
 export { createIfNotExists } from "./createIfNotExists";
+export { getDirectories } from "./getDirectories";
 export { exists } from "./exists";
+export { del } from "./del";