+++ packages/cli-forger/src/workers/prompt/ask.ts
@@ -1,25 +1,3 @@
-import inquirer from "inquirer";
-import {
-  Answers,
-  Question,
-  QuestionAnswerMap,
-  QuestionArray,
-  QuestionObservable,
-} from "inquirer/dist/cjs/types/types";
-import { workers } from "..";
+import { prompt } from "enquirer";
 
-export async function ask<T extends Answers>(
-  questions:
-    | QuestionArray<T>
-    | QuestionAnswerMap<T>
-    | QuestionObservable<T>
-    | Question<T>,
-  answers?: Partial<T>,
-) {
-  try {
-    const res = await inquirer.prompt(questions, answers);
-    return res;
-  } catch {
-    workers.logger.exit.error("Canceled");
-  }
-}
+export { prompt as ask };