+++ packages/filegen-cli/src/commands/Init.ts
@@ -1,9 +1,0 @@
-import { GluegunToolbox } from 'gluegun'
-
-module.exports = {
-  name: 'init',
-
-  run: async (toolbox: GluegunToolbox) => {
-    await toolbox.createFile('.', 'filegen.ts', 'initial.ts.ejs')
-  },
-}