+++ packages/filegen-cli/src/index.ts
@@ -1,0 +1,8 @@
+#!/usr/bin/env node
+"use strict";
+
+export { type Config } from "./types";
+
+import cli from "./cli";
+
+cli.run(process.argv);