+++ packages/filegen-cli/src/infra/reader/contracts/Reader.ts
@@ -1,3 +1,0 @@
-export abstract class Reader {
-  abstract findConfigFile(): string
-}