+++ packages/filegen-cli/src/shared/errors/NotFoundError.ts
@@ -1,7 +1,0 @@
-import { ApplicationError } from "../core/errors/ApplicationError";
-
-export class NotFoundError extends ApplicationError {
-  constructor(notFoundMessage: string) {
-    super(`${notFoundMessage} not found`)
-  }
-}